requirements.txt
flask
tensorflow-cpu==2.12.0
keras==2.14.0
opencv-python
numpy
pandas
matplotlib
scikit-learn